pvs: fixed warnings for fatfs timestamp

+++ firmware/targets/f7/fatfs/fatfs.c
@@ -20,6 +20,7 @@ DWORD get_fattime() {
     FuriHalRtcDateTime furi_time;
     furi_hal_rtc_get_datetime(&furi_time);
 
-    return (furi_time.year - 1980) << 25 | furi_time.month << 21 | furi_time.day << 16 |
-           furi_time.hour << 11 | furi_time.minute << 5 | furi_time.second >> 1;
+    return ((uint32_t)(furi_time.year - 1980) << 25) | furi_time.month << 21 |
+           furi_time.day << 16 | furi_time.hour << 11 | furi_time.minute << 5 |
+           furi_time.second >> 1;
 }
